+++ tests/Integration/Dto/Customer/CustomerTest.php
@@ -34,37 +34,6 @@ use Shopgate\ConnectSdk\Tests\Integration\CustomerTest as CustomerBaseTest;
 
 class CustomerTest extends CustomerBaseTest
 {
-    const CUSTOMER_CUSTOMER_EXTERNAL_CUSTOMER_CODE = 'external_customer_code';
-    const CUSTOMER_CUSTOMER_FIRSTNAME = 'Firstname';
-    const CUSTOMER_CUSTOMER_LASTNAME = 'Lastname';
-    const CUSTOMER_CUSTOMER_MIDDLE_NAME = 'Middlename';
-    const CUSTOMER_CUSTOMER_EMAIL = 'example+%s@mail.com';
-    const CUSTOMER_SETTINGS_DEFAULT_LOCALE = 'en-us';
-    const CUSTOMER_SETTINGS_DEFAULT_CURRENCY = 'USD';
-    const CUSTOMER_SETTINGS_DEFAULT_COMMUNICATION_PREFERENCES = ['email', 'sms'];
-    const CUSTOMER_SETTINGS_DEFAULT_LOCATION_CODE = 'DERetail001';
-    const CUSTOMER_ATTRIBUTE_CODE = 'attribute_code';
-    const CUSTOMER_ATTRIBUTE_NAME = 'Attribute name';
-    const CUSTOMER_ATTRIBUTE_VALUE_CODE = 'attribute_value_code';
-    const CUSTOMER_ATTRIBUTE_VALUE_NAME = 'Attribute Value Name';
-    const CUSTOMER_CONTACT_EXTERNAL_CUSTOMER_CODE = 'customer_code';
-    const CUSTOMER_CONTACT_FIRSTNAME = 'Firstname';
-    const CUSTOMER_CONTACT_MIDDLE_NAME = 'Middlename';
-    const CUSTOMER_CONTACT_LAST_NAME = 'Lastname';
-    const CUSTOMER_CONTACT_COMPANY = 'Shopgate Inc';
-    const CUSTOMER_CONTACT_ADDRESS_1 = 'Somestreet 12';
-    const CUSTOMER_CONTACT_ADDRESS_2 = 'Address 2';
-    const CUSTOMER_CONTACT_ADDRESS_3 = 'Address 3';
-    const CUSTOMER_CONTACT_ADDRESS_4 = 'Address 4';
-    const CUSTOMER_CONTACT_CITY = 'Austin';
-    const CUSTOMER_CONTACT_POSTAL_CODE = '78732';
-    const CUSTOMER_CONTACT_REGION = 'TX';
-    const CUSTOMER_CONTACT_COUNTRY = 'US';
-    const CUSTOMER_CONTACT_PHONE = '+1000000001';
-    const CUSTOMER_CONTACT_FAX = '+1000000002';
-    const CUSTOMER_CONTACT_MOBILE = '+1000000003';
-    const CUSTOMER_CONTACT_EMAIL = 'somelocation+%s@someRetailer.com';
-
     /**
      * @throws Exception
      */
@@ -88,41 +57,6 @@ class CustomerTest extends CustomerBaseTest
         // Assert
         /** @noinspection PhpParamsInspection */
         $this->assertCount($createdItemCount, $customers);
-    }
-
-    /**
-     * @param int $itemCount
-     *
-     * @return Customer\Create[]
-     *
-     * @throws Exception
-     */
-    private function provideSampleCustomers($itemCount = 2)
-    {
-        $result = [];
-        for ($count = 1; $count < ($itemCount + 1); $count++) {
-            $customer = new Customer\Create();
-            $customer->setExternalCustomerNumber(self::CUSTOMER_CUSTOMER_EXTERNAL_CUSTOMER_CODE . $count);
-            $customer->setFirstName(self::CUSTOMER_CUSTOMER_FIRSTNAME . $count);
-            $customer->setMiddleName(self::CUSTOMER_CUSTOMER_MIDDLE_NAME . $count);
-            $customer->setLastName(self::CUSTOMER_CUSTOMER_LASTNAME . $count);
-            $customer->setEmailAddress(sprintf(self::CUSTOMER_CUSTOMER_EMAIL, $count));
-            $customer->setStatus(Customer\Create::STATUS_ACTIVE);
-            $customer->setIsAnonymous(false);
-
-            $setting = new Customer\Dto\Settings();
-            $setting->setDefaultLocale('en-us');
-            $setting->setDefaultCurrency('USD');
-            $setting->setCommunicationPreferences(['email']);
-            $setting->setDefaultLocationCode('DERetail001');
-            $setting->setMarketingOptIn(true);
-
-            $customer->setSettings($setting);
-
-            $result[] = $customer;
-        }
-
-        return $result;
     }
 
     /**
@@ -173,7 +107,7 @@ class CustomerTest extends CustomerBaseTest
         // Assert
         /** @noinspection PhpParamsInspection */
         $this->assertEquals(
-            self::CUSTOMER_CUSTOMER_EXTERNAL_CUSTOMER_CODE . '2',
+            self::CUSTOMER_EXTERNAL_CUSTOMER_CODE . '2',
             $customer->getExternalCustomerNumber()
         );
     }
@@ -210,11 +144,11 @@ class CustomerTest extends CustomerBaseTest
     {
         // Arrange
         $customer = new Customer\Create();
-        $customer->setExternalCustomerNumber(self::CUSTOMER_CUSTOMER_EXTERNAL_CUSTOMER_CODE);
-        $customer->setFirstName(self::CUSTOMER_CUSTOMER_FIRSTNAME);
-        $customer->setMiddleName(self::CUSTOMER_CUSTOMER_MIDDLE_NAME);
-        $customer->setLastName(self::CUSTOMER_CUSTOMER_LASTNAME);
-        $customer->setEmailAddress(sprintf(self::CUSTOMER_CUSTOMER_EMAIL, 1));
+        $customer->setExternalCustomerNumber(self::CUSTOMER_EXTERNAL_CUSTOMER_CODE);
+        $customer->setFirstName(self::CUSTOMER_FIRSTNAME);
+        $customer->setMiddleName(self::CUSTOMER_MIDDLE_NAME);
+        $customer->setLastName(self::CUSTOMER_LASTNAME);
+        $customer->setEmailAddress(sprintf(self::CUSTOMER_EMAIL, 1));
         $customer->setStatus(Customer\Create::STATUS_ACTIVE);
         $customer->setIsAnonymous(false);
 
@@ -304,10 +238,10 @@ class CustomerTest extends CustomerBaseTest
         /** @noinspection PhpParamsInspection */
         // General
         $this->assertEquals(Customer\Create::STATUS_ACTIVE, $customer->getStatus());
-        $this->assertEquals(self::CUSTOMER_CUSTOMER_FIRSTNAME, $customer->getFirstName());
-        $this->assertEquals(self::CUSTOMER_CUSTOMER_LASTNAME, $customer->getLastName());
-        $this->assertEquals(self::CUSTOMER_CUSTOMER_MIDDLE_NAME, $customer->getMiddleName());
-        $this->assertEquals(sprintf(self::CUSTOMER_CUSTOMER_EMAIL, 1), $customer->getEmailAddress());
+        $this->assertEquals(self::CUSTOMER_FIRSTNAME, $customer->getFirstName());
+        $this->assertEquals(self::CUSTOMER_LASTNAME, $customer->getLastName());
+        $this->assertEquals(self::CUSTOMER_MIDDLE_NAME, $customer->getMiddleName());
+        $this->assertEquals(sprintf(self::CUSTOMER_EMAIL, 1), $customer->getEmailAddress());
         $this->assertFalse($customer->getIsAnonymous());
 
         // Settings
@@ -392,10 +326,10 @@ class CustomerTest extends CustomerBaseTest
         $response = $this->sdk->getCustomerService()->addCustomers($sampleCustomers);
 
         $customerUpdate = new Customer\Update();
-        $customerUpdate->setFirstName(self::CUSTOMER_CUSTOMER_FIRSTNAME . ' Update');
-        $customerUpdate->setMiddleName(self::CUSTOMER_CUSTOMER_MIDDLE_NAME . ' Update');
-        $customerUpdate->setLastName(self::CUSTOMER_CUSTOMER_LASTNAME . ' Update');
-        $customerUpdate->setEmailAddress(sprintf(self::CUSTOMER_CUSTOMER_EMAIL, 'update'));
+        $customerUpdate->setFirstName(self::CUSTOMER_FIRSTNAME . ' Update');
+        $customerUpdate->setMiddleName(self::CUSTOMER_MIDDLE_NAME . ' Update');
+        $customerUpdate->setLastName(self::CUSTOMER_LASTNAME . ' Update');
+        $customerUpdate->setEmailAddress(sprintf(self::CUSTOMER_EMAIL, 'update'));
         $customerUpdate->setStatus(Customer\Create::STATUS_INACTIVE);
         $customerUpdate->setIsAnonymous(true);
 
@@ -426,10 +360,10 @@ class CustomerTest extends CustomerBaseTest
 
         // General
         $this->assertEquals(Customer\Create::STATUS_INACTIVE, $customer->getStatus());
-        $this->assertEquals(self::CUSTOMER_CUSTOMER_FIRSTNAME . ' Update', $customer->getFirstName());
-        $this->assertEquals(self::CUSTOMER_CUSTOMER_LASTNAME . ' Update', $customer->getLastName());
-        $this->assertEquals(self::CUSTOMER_CUSTOMER_MIDDLE_NAME . ' Update', $customer->getMiddleName());
-        $this->assertEquals(sprintf(self::CUSTOMER_CUSTOMER_EMAIL, 'update'), $customer->getEmailAddress());
+        $this->assertEquals(self::CUSTOMER_FIRSTNAME . ' Update', $customer->getFirstName());
+        $this->assertEquals(self::CUSTOMER_LASTNAME . ' Update', $customer->getLastName());
+        $this->assertEquals(self::CUSTOMER_MIDDLE_NAME . ' Update', $customer->getMiddleName());
+        $this->assertEquals(sprintf(self::CUSTOMER_EMAIL, 'update'), $customer->getEmailAddress());
         $this->assertTrue($customer->getIsAnonymous());
 
         // Settings